add delete task runner button
